take comments into consideration
