+++ web/vite.config.ts
@@ -2,7 +2,7 @@ import { resolve } from "node:path";
 import tanstackRouter from "@tanstack/router-plugin/vite";
 import viteReact from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
-import { defineConfig } from "vitest/config";
+import { defineConfig } from "vite";
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -15,9 +15,5 @@ export default defineConfig({
     alias: {
       "@": resolve(import.meta.dirname, "./src"),
     },
-  },
-  test: {
-    environment: "node",
-    include: ["src/**/*.test.{ts,tsx}"],
   },
 });
